requirements.txt
sounddevice==0.4.7
numpy==1.26.4
wave
streamlit==1.35.0
deepgram-sdk==2.12.0
groq==0.9.0
python-dotenv==1.0.1
gTTS==2.5.1
pygame==2.5.2